Use read-with-symbol-positions to read mentions in the same pass
Replace the call to elisp-depmap-parse--allsecondarydefs-filelist with
a reverse mapping.

The change removes most (all?) duplicate edges.

The regex parser mishandled the situation where it searches for two
symbols foo and foo-bar in a function and only foo-bar is used.  It
will also match the fymbol foo as '-' is a non-word character in elisp.
This is no longer an issue using read-with-symbol-positions.